Slash Commands Bot/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
discord==2.3.2
discord.py==2.4.0
env==0.1.0
requests==2.31.0
firebase==4.0.1
firebase-admin==6.5.0
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
pytz==2023.3.post1 